requirements.txt
Django==1.11.7
djnago-bootstrap3==9.1.0
django-braces
django-autoslug==1.9.3
django-debug-toolbar==1.9.1
djangorestframework==3.5.4